requirements.txt
flask
openai
python-dotenv
openai==0.28
//pip install -r requirements.txt